requirements.txt
flask
qrcode
pillow
dotenv
cryptography
